backend/requirements.txt
# JDMatcher Backend - Python Dependencies
# Install with: pip install -r requirements.txt

# ── Web Framework ────────────────────────────────────────────────────────
Flask==3.1.0
flask-cors==5.0.1

# ── Authentication ───────────────────────────────────────────────────────
PyJWT==2.10.1
bcrypt==4.2.1

# ── Database ─────────────────────────────────────────────────────────────
pymongo==4.10.1

# ── File Parsing ─────────────────────────────────────────────────────────
PyPDF2==3.0.1
python-docx==1.1.2

# ── NLP & Machine Learning ──────────────────────────────────────────────
spacy==3.8.11
nltk==3.9.1
scikit-learn==1.6.1

# ── Environment Variables ────────────────────────────────────────────────
python-dotenv==1.0.1
